ecovision-ai-nexus/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.12
numpy==2.1.3
pillow==11.0.0
